+++ crates/pctx_code_execution_runtime/Cargo.toml
@@ -13,6 +13,7 @@ categories = ["development-tools"]
 path = "src/lib.rs"
 
 [dependencies]
+pctx_registry = { version = "^0.1.0", path = "../pctx_registry" }
 pctx_config = { version = "^0.1.3", path = "../pctx_config" }
 deno_core = { workspace = true }
 serde = { workspace = true }
