+++ crates/pctx_code_execution_runtime/src/callback_ops.rs
@@ -7,7 +7,7 @@ use deno_core::{OpState, op2};
 use std::cell::RefCell;
 use std::rc::Rc;
 
-use crate::{CallbackRegistry, error::McpError};
+use pctx_registry::{CallbackRegistry, RegistryError};
 
 #[op2(async)]
 #[serde]
@@ -15,7 +15,7 @@ pub(crate) async fn op_invoke_callback(
     state: Rc<RefCell<OpState>>,
     #[string] id: String,
     #[serde] arguments: Option<serde_json::Value>,
-) -> Result<serde_json::Value, McpError> {
+) -> Result<serde_json::Value, RegistryError> {
     let registry = {
         let borrowed = state.borrow();
         borrowed.borrow::<CallbackRegistry>().clone()
