+++ crates/pctx_executor/Cargo.toml
@@ -18,6 +18,7 @@ node_resolver = { workspace = true }
 sys_traits = { workspace = true }
 pctx_deno_transpiler = { version = "^0.1.1", path = "../pctx_deno_transpiler" }
 pctx_code_execution_runtime = { version = "^0.1.3", path = "../pctx_code_execution_runtime" }
+pctx_registry = { version = "^0.1.0", path = "../pctx_registry" }
 pctx_type_check_runtime = { version = "^0.1.2", path = "../pctx_type_check_runtime" }
 pctx_config = { version = "^0.1.3", path = "../pctx_config" }
 tempfile = "3"
