+++ crates/pctx_session_server/Cargo.toml
@@ -14,7 +14,7 @@ path = "src/bin/generate-openapi.rs"
 # Local crates
 pctx_code_mode = { path = "../pctx_code_mode" }
 pctx_config = { path = "../pctx_config" }
-pctx_code_execution_runtime = { path = "../pctx_code_execution_runtime" }
+pctx_registry = { path = "../pctx_registry" }
 pctx_codegen = { path = "../pctx_codegen" }
 
 # Web framework
